+++ projects/md-dynamic-forms-core/src/lib/model/form/array/field-table.ts
@@ -1,17 +1,21 @@
 import {TableConfig} from '../../table/table-config.interface';
 import {FieldArray, FieldArrayConfig} from './field-array';
+import {TableColumnConfig} from '../../table/table-column-config.interface';
 
-export interface FieldTableConfig extends FieldArrayConfig {
-  config: TableConfig;
+export interface FieldTableConfig<T extends TableConfig, TC extends TableColumnConfig> extends FieldArrayConfig {
+  config: T;
+  columns: TC[];
 }
 
-export class FieldTable extends FieldArray {
-  config: TableConfig;
+export class FieldTable<T extends TableConfig, TC extends TableColumnConfig> extends FieldArray {
+  config: T;
+  columns: TC[];
 
-  constructor(config: FieldTableConfig) {
+  constructor(config: FieldTableConfig<T, TC>) {
     super(config);
 
     this.config = config.config;
+    this.columns = config.columns;
     this.formType = 'array';
     this.component = 'table';
   }