+++ projects/md-dynamic-forms-core/src/lib/model/table/table-column-config.interface.ts
@@ -3,7 +3,4 @@ export interface TableColumnConfig {
   width: number;
   name: string;
   sortable?: boolean;
-  component?: any;
-  fieldtype?: string;
-  readonly?: boolean;
 }